requirements.txt
autopep8==2.0.2
boto3==1.26.99
botocore==1.29.99
click==8.1.3
Flask==2.2.3
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.2
pycodestyle==2.10.0
python-dateutil==2.8.2
python-dotenv==1.0.0
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.15
uWSGI==2.0.21
Werkzeug==2.2.3
